requirements.txt
uvicorn==0.30.1
gunicorn==22.0.0
pytest==8.2.2
django==5.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
fontawesomefree==6.5.1
loguru==0.7.2
requests==2.32.3
Pillow==10.3.0
bs4==0.0.2
django-meta==2.4.2
django-select2==8.1.2
elasticsearch-dsl==8.13.1
psycopg==3.1.19
psycopg-binary==3.1.19
wikipedia==1.4.0
python-decouple==3.8
pytesseract==0.3.10
pdfminer.six==20231228
pdf2image==1.17.0
PyPDF2==3.0.1
pydantic==2.8.0
celery==5.4.0
redis==5.0.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
drf-redesign==0.3.9
django-ckeditor==6.7.1
django-role-permissions==3.2.0
django-meta